user_leds: split gpio.c
